requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.1.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
spacy>=3.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
